Implementing correct handling of ICE messages
Previous to this PR, only the very first ICE message has been transferred between
the initiator and the follower. This made WebRTC only work in a local network.

To send multiple ICE messages, the current structure of the libraries has been
deemed inappropriate. So all of it has been changed to use channels between
each layer of abstraction. To handle all messages, a 'process' method
in each layer of abstraction needs to be called in regular intervals.